refactor: Extract script functionality
